requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
flask_cors==3.0.8
flask-marshmallow==0.14.0
psycopg2-binary==2.8.6
zappa==0.52.0
